Bluetooth: Fix LE connection interval type

+++ net/bluetooth/conn_internal.h
@@ -37,7 +37,7 @@ struct bt_conn_le {
 
 	bt_addr_le_t		init_addr;
 	bt_addr_le_t		resp_addr;
-	uint8_t			conn_interval;
+	uint16_t		conn_interval;
 	uint8_t			features[8];
 };
